Rename array-methods.js to collection-methods.js
