+++ lib/collection-methods.js
