+++ lib/index.js
@@ -18,7 +18,7 @@
 
 'use strict';
 
-var arrayMethods = require('./array-methods');
+var arrayMethods = require('./collection-methods');
 var realmConstructor;
 
 if (typeof Realm != 'undefined') {
